some architecture refactor
