+++ molina/__init__.py
@@ -1,6 +1,6 @@
 ''''''
 
-from molina.ui import MainWindow
+from molina.main_window import MainWindow
 
 __version__ = '0.0.1'
